# api/openapi.yaml
openapi: 3.0.3
info:
  title: ASM Platform API
  version: 0.8.0
  description: |
    Contract for the HTTP routes currently mounted by cmd/api. Discovery scope,
    task template, and task run management are implemented below service level
    but are not listed here until HTTP handlers are added.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects/{project_id}/reports/dashboard:
    get:
      tags: [reports]
      summary: Get dashboard summary.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Dashboard data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /projects/{project_id}/reports/trends:
    get:
      tags: [reports]
      summary: Get trend series.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Trend series.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /projects/{project_id}/reports/top:
    get:
      tags: [reports]
      summary: Get top-N report.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: dimension
          in: query
          schema: { type: string, default: asset }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Top-N data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /projects/{project_id}/reports/remediation:
    get:
      tags: [reports]
      summary: Get remediation summary.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Remediation data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /projects/{project_id}/reports/exports:
    get:
      tags: [reports]
      summary: List report export jobs.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/PageNumber"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Export jobs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
    post:
      tags: [reports]
      summary: Create an async report export.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "201":
          description: Export job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /projects/{project_id}/reports/exports/{id}:
    get:
      tags: [reports]
      summary: Get export job detail.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Export job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /projects/{project_id}/reports/exports/{id}/download:
    get:
      tags: [reports]
      summary: Download a completed export file.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Export file stream.
          content:
            text/csv: { schema: { type: string, format: binary } }
            application/json: { schema: { type: string, format: binary } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: integer, format: uint64, minimum: 1 }
...
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: uint64, minimum: 1 }
...
    PageNumber:
      name: page_number
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
    GenericEnvelope:
      type: object
      required: [request_id, data]
      properties:
        request_id: { type: string }
        data: {}
...